# Adding SIMD AVX Vectorization optimization function to Cyclic EVD