--- a/src/evd/cyclic/CMakeLists.txt
+++ b/src/evd/cyclic/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(SOURCES
     evd_cyclic.cpp
+    evd_cyclic_vectorize.cpp
     evd_cyclic.hpp
     )