--- a/test/evd/cyclic/CMakeLists.txt
+++ b/test/evd/cyclic/CMakeLists.txt
@@ -1,8 +1,11 @@
 add_executable(evd_cyclic_test evd_cyclic_test.cpp)
+add_executable(evd_cyclic_vectorize_test evd_cyclic_vectorize_test.cpp)
 add_executable(evd_cyclic_tol_test evd_cyclic_tol_test.cpp)
 
 target_link_libraries(evd_cyclic_test gtest_main evd-cyclic-lib utils-lib m)
+target_link_libraries(evd_cyclic_vectorize_test gtest_main evd-cyclic-lib utils-lib m)
 target_link_libraries(evd_cyclic_tol_test gtest_main evd-cyclic-lib utils-lib m)
 
 add_test(NAME EVD_cyclic COMMAND evd_cyclic_test)
+add_test(NAME evd_cyclic_vectorize COMMAND evd_cyclic_vectorize_test)
 add_test(NAME EVD_cyclic_tol COMMAND evd_cyclic_tol_test)