This PR introduces AVX intrinsics into Cyclic EVD. The performance improvement is quite satisfactory (more than double for matrix size 64 : I have verified the results from matrix sizes from 8 to 64)

It's quite straightforward to understand, where 4 double precision values are loaded, the arithmetic operations are performed on the vector and stored back into memory. In case of column operations, the loading happens via `_mm256_set_pd` intrinsic, which is much slower than general loads since the memory locations are not contiguous. As for the row operations, aligned intrinsics don't seem to work as we get segmentation fault. I have used unaligned ones, and that can be investigated further I think.

I guess it can be still optimized though (I need to try using FMA), as more options still need to be checked. The cost function has not been altered since the total number of operations remain the same (needs some validation regarding the flop count).

Results (on my machine, Core i5 6200 U @ 2.3 GHz - 2.8 GHz max. and turbo boost was not disabled)

```
Performance benchmark on array of size 8 by 8
evd_cyclic
Runtime:     469113 cycles (median)
Performance: 0.954995 flops/cycle (median)
----------

evd_cyclic_vectorize
Runtime:     424345 cycles (median)
Performance: 1.05575 flops/cycle (median)
----------

Performance benchmark on array of size 16 by 16
evd_cyclic
Runtime:     2.5936e+06 cycles (median)
Performance: 1.40654 flops/cycle (median)
----------

evd_cyclic_vectorize
Runtime:     1.98335e+06 cycles (median)
Performance: 1.83932 flops/cycle (median)
----------

Performance benchmark on array of size 32 by 32
evd_cyclic
Runtime:     1.60741e+07 cycles (median)
Performance: 1.82674 flops/cycle (median)
----------

evd_cyclic_vectorize
Runtime:     1.08304e+07 cycles (median)
Performance: 2.71118 flops/cycle (median)
----------

Performance benchmark on array of size 64 by 64
evd_cyclic
Runtime:     1.62696e+08 cycles (median)
Performance: 1.44729 flops/cycle (median)
----------

evd_cyclic_vectorize
Runtime:     7.69475e+07 cycles (median)
Performance: 3.06012 flops/cycle (median)
----------
```