@RajaSimha 2019-05-10
> Thanks for the initial implementation. Could you create a new `.cpp` file for the implementation. I think it is okay to use only one `.hpp` file for all EVD cyclic variants. Further, could you create a new test file that will be basically the copy of evd cyclic test file which calls this new function? This way we will know that the new version returns the same results.

Done!